Global StyleSheet & fixed issue #8
