+++ client/app/screens/TrailScreen.tsx
@@ -20,17 +20,14 @@ import MapView, { Polyline } from "react-native-maps";
 import {
   ActivityIndicator,
   Alert,
-  Dimensions,
-  Platform,
-  StyleSheet,
   Text,
-  TouchableOpacity,
   View,
 } from "react-native";
 import MapButton from "../components/MapButton";
 import SaveTrailModal from "../components/SaveTrailModal";
 import { SafeAreaView } from "react-native-safe-area-context";
 import LoginButton from "../components/LoginButton";
+import styles from  "../components/Styles";
 
 // Constants
 import { CAMP_ALLEN_COORDS } from "../utils/constants";
@@ -280,12 +277,20 @@ export default function TrailScreen({ navigation, route }: TrailScreenProps) {
         )}
       </MapView>
 
-      {/* login button is relative to map */}
+      {/* Login button is relative to map */}
       <View style={[styles.loginBtnContainer]}>
         {!auth.isAuthenticated && (
           <LoginButton onPress={() => navigation.navigate("Admin")} />
         )}
       </View>
+
+      {/* Update Password */}
+        {userId ? (
+          <View style={[styles.loginBtnContainer]}>
+            <LoginButton onPress={() => navigation.navigate("Update Password")}/>
+          </View>): (
+          <></>
+        )}
 
       {/* Other button containers are at the bottom of the screen */}
       <View style={[styles.fgContainer]}>
@@ -356,14 +361,6 @@ export default function TrailScreen({ navigation, route }: TrailScreenProps) {
           )}
         </View>
 
-        {/*Change Password*/}
-        {userId ? (
-          <View style={[styles.loginBtnContainer]}>
-            <LoginButton onPress={() => navigation.navigate("Update Password")}/>
-          </View>): (
-          <></>
-        )}
-
         {/* Show these buttons for a logged in user */}
         {userId ? (
           <View style={styles.btnContainer}>
@@ -415,53 +412,3 @@ export default function TrailScreen({ navigation, route }: TrailScreenProps) {
     </SafeAreaView>
   );
 }
-
-const styles = StyleSheet.create({
-  safeAreaView: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-  },
-
-  fgContainer: {
-    position: "absolute",
-    top: 0,
-    left: 0,
-    // todo: figure out the proper way to account for header
-    height: "100%",
-    width: "100%",
-
-    flex: 1,
-    alignItems: "center",
-    justifyContent: "flex-end",
-  },
-
-  map: {
-    width: Dimensions.get("window").width,
-    height: Dimensions.get("window").height,
-  },
-
-  loginBtnContainer: {
-    position: "absolute",
-
-    right: 10,
-    top: Platform.OS === "ios" ? 35 : 110,
-
-    zIndex: 3, // for iOS
-    elevation: 3, // for Android
-  },
-
-  btnContainer: {
-    flexDirection: "row",
-    justifyContent: "space-around",
-    width: "100%",
-    margin: 10,
-  },
-
-  permissionsText: {
-    padding: 10,
-    color: "red",
-    fontSize: 24,
-    textAlign: "center",
-  },
-});
